backend/requirements.txt
aiomysql==0.2.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==24.2
passlib==1.7.4
pdf2image==1.17.0
pillow==11.1.0
pluggy==1.5.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyMySQL==1.1.1
pytest==8.3.5
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
